% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_pollen_levels.R
\name{compare_pollen_levels}
\alias{compare_pollen_levels}
\title{Compare pollen levels between two cities}
\usage{
compare_pollen_levels(
  city1_name,
  city2_name,
  city1_lat,
  city1_lon,
  city2_lat,
  city2_lon
)
}
\arguments{
\item{city1_name}{Name of the first city}

\item{city2_name}{Name of the second city}

\item{city1_lat}{Latitude of the first city}

\item{city1_lon}{Longitude of the first city}

\item{city2_lat}{Latitude of the second city}

\item{city2_lon}{Longitude of the second city}
}
\value{
A string with the name of the city that has the least pollen
}
\description{
Compare pollen levels between two cities
}
