% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_pollen_forecast.R
\name{get_pollen_forecast}
\alias{get_pollen_forecast}
\title{Getting the pollen data from the API}
\usage{
get_pollen_forecast(latitude, longitude)
}
\arguments{
\item{latitude}{Latitude of the location}

\item{longitude}{Longitude of the location}
}
\value{
A data.frame with pollen forecast data
}
\description{
Getting the pollen data from the API
}
\examples{
df_test <- get_pollen_forecast(latitude=46.94, longitude=7.94)

}
