% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_pollen_with_cache.R
\name{get_pollen_forecast_with_cache}
\alias{get_pollen_forecast_with_cache}
\title{Read data from the API, uses caching}
\usage{
get_pollen_forecast_with_cache(latitude, longitude)
}
\arguments{
\item{latitude}{Latitude of the location}

\item{longitude}{Longitude of the location}
}
\value{
A data.frame with pollen forecast data
}
\description{
Checks if a file with the requested data exists, if so it uses that file to return the requested data,
otherwise it makes an API call.
}
\examples{
latitude <- 47.3769
longitude <- 8.5417

df_test <- get_pollen_forecast_with_cache(latitude, longitude)

}
